third_party/blink/web_tests/platform/generic/editing/pasteboard/copy-in-password-field-expected.txt

This test verifies that copying is disabled for password fields by attempting to copy from a password field and paste into a textfield. If the test passes, you'll see a of 'PASS' message below, and the textfield will remain unmodified.

(The test is only meaningful in debug builds, since paste is disabled in release builds.)

 

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("text").value is "[not modified]"
